fix: compile with GCC 15

+++ src/common/Typedefs.hpp
@@ -22,15 +22,15 @@ using s16 = std::int16_t;
 using s32 = std::int32_t;
 using s64 = std::int64_t;
 
-constexpr u8  operator""_u8 (u64 const nbr) { return static_cast<u8> (nbr); }
-constexpr u16 operator""_u16(u64 const nbr) { return static_cast<u16>(nbr); }
-constexpr u32 operator""_u32(u64 const nbr) { return static_cast<u32>(nbr); }
-constexpr u64 operator""_u64(u64 const nbr) { return nbr;                   }
+constexpr u8  operator""_u8 (unsigned long long const nbr) { return static_cast<u8> (nbr); }
+constexpr u16 operator""_u16(unsigned long long const nbr) { return static_cast<u16>(nbr); }
+constexpr u32 operator""_u32(unsigned long long const nbr) { return static_cast<u32>(nbr); }
+constexpr u64 operator""_u64(unsigned long long const nbr) { return nbr;                   }
 
-constexpr s8  operator""_s8 (u64 const nbr) { return static_cast<s8> (nbr); }
-constexpr s16 operator""_s16(u64 const nbr) { return static_cast<s16>(nbr); }
-constexpr s32 operator""_s32(u64 const nbr) { return static_cast<s32>(nbr); }
-constexpr s64 operator""_s64(u64 const nbr) { return static_cast<s64>(nbr); }
+constexpr s8  operator""_s8 (unsigned long long const nbr) { return static_cast<s8> (nbr); }
+constexpr s16 operator""_s16(unsigned long long const nbr) { return static_cast<s16>(nbr); }
+constexpr s32 operator""_s32(unsigned long long const nbr) { return static_cast<s32>(nbr); }
+constexpr s64 operator""_s64(unsigned long long const nbr) { return static_cast<s64>(nbr); }
 
 template<typename T>
 using Opt = std::optional<T>;
